+++ ui/src/components/app-user/QasAppUser.vue
@@ -226,6 +226,7 @@ function useAvatarNotifications () {
     const classes = [
       'qas-app-user__notification-avatar',
       'animated',
+      'absolute',
       {
         rubberBand: hasAnimated.value
       }